# docs/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Submission API",
    "version": "1.0.0",
    "description": "API for managing code submissions"
  },
  "servers": [
    {
      "url": "https://codenire.io"
    }
  ],
  "paths": {
    "/run": {
      "post": {
        "summary": "Run Multi Files Submission",
        "operationId": "runFilesSubmission",
        "tags": [
          "Submission"
        ],
        "requestBody": {
          "description": "Run Files Submission",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission ran successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/session/start": {
      "post": {
        "summary": "Start Code Playground Long Living Session",
        "operationId": "startSession",
        "tags": [
          "Submission"
        ],
        "requestBody": {
          "description": "Start Code Playground Long Living Session",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created successfully"
          }
        }
      }
    },
    "/actions": {
      "get": {
        "summary": "Get with refresh Action List",
        "operationId": "refreshActionList",
        "tags": [
          "Action"
        ],
        "responses": {
          "200": {
            "description": "Get with refresh Action List",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionListResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Get Actual Action List",
        "operationId": "getActionList",
        "tags": [
          "Action"
        ],
        "responses": {
          "200": {
            "description": "Get with refresh Action List",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionListResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CommonSubmissionRequest": {
        "type": "object",
        "properties": {
          "TemplateId": {
            "type": "string"
          },
          "Args": {
            "type": "string"
          },
          "Stdin": {
            "description": "data which will available via stdin reader",
            "type": "string"
          },
          "ActionId": {
            "type": "string",
            "default": "default"
          },
          "ExternalOptions": {
            "type": "object",
            "description": "external options like CompileCmd or RunCmd",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "TemplateId",
          "Args",
          "Stdin"
        ]
      },
      "StartSessionRequest": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "Version": {
            "type": "integer"
          },
          "Cluster": {
            "type": "integer"
          }
        },
        "required": [
          "Id"
        ]
      },
      "StartSessionResponse": {
        "type": "object",
        "properties": {
          "Version": {
            "type": "integer"
          },
          "Cluster": {
            "type": "integer"
          }
        }
      },
      "SubmissionRequest": {
        "type": "object",
        "allOf": [
          {
            "type": "object",
            "properties": {
              "Files": {
                "type": "object",
                "example": {
                  "index.js": "...some content",
                  "src/script.js": "...some content",
                  "foo/bar/script.js": "...some content",
                  "package.json": "...some content"
                },
                "additionalProperties": {
                  "type": "string"
                }
              }
            },
            "required": [
              "Files"
            ]
          },
          {
            "$ref": "#/components/schemas/CommonSubmissionRequest"
          }
        ]
      },
      "SubmissionScriptRequest": {
        "type": "object",
        "allOf": [
          {
            "type": "object",
            "properties": {
              "Code": {
                "type": "string"
              }
            },
            "required": [
              "Code"
            ]
          },
          {
            "$ref": "#/components/schemas/CommonSubmissionRequest"
          }
        ]
      },
      "SubmissionResponse": {
        "type": "object",
        "properties": {
          "Events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmissionResponseEvents"
            }
          },
          "RunEnvironment": {
            "$ref": "#/components/schemas/RunEnvironment"
          }
        },
        "required": [
          "Events",
          "RunEnvironment"
        ]
      },
      "SubmissionResponseEvents": {
        "type": "object",
        "properties": {
          "Message": {
            "type": "string"
          },
          "Kind": {
            "type": "string"
          }
        },
        "required": [
          "Message",
          "Kind"
        ]
      },
      "RunEnvironment": {
        "type": "object",
        "properties": {
          "CompileCmd": {
            "type": "string"
          },
          "RunCmd": {
            "type": "string"
          },
          "RunTime": {
            "type": "number"
          },
          "CompileTime": {
            "type": "number"
          },
          "ActionName": {
            "type": "string"
          }
        },
        "required": [
          "CompileCmd",
          "RunCmd",
          "RunTime",
          "CompileTime",
          "ActionName"
        ]
      },
      "SandboxResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "exitCode": {
            "type": "integer"
          },
          "stdout": {
            "type": "string",
            "format": "byte"
          },
          "stderr": {
            "type": "string",
            "format": "byte"
          },
          "RunEnvironment": {
            "$ref": "#/components/schemas/RunEnvironment"
          }
        },
        "required": [
          "exitCode",
          "stdout",
          "stderr",
          "RunEnvironment"
        ]
      },
      "SandboxRequest": {
        "type": "object",
        "properties": {
          "sandId": {
            "type": "string"
          },
          "binary": {
            "type": "string",
            "description": "files in tar archive encoded with base64"
          },
          "args": {
            "type": "string"
          },
          "extendedOptions": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "stdin": {
            "description": "data which will available via stdin reader",
            "type": "string"
          },
          "action": {
            "type": "string"
          }
        },
        "required": [
          "sandId",
          "binary",
          "args",
          "stdin",
          "action"
        ]
      },
      "ContainerOptions": {
        "type": "object",
        "properties": {
          "CompileTTL": {
            "type": "integer"
          },
          "RunTTL": {
            "type": "integer"
          },
          "MemoryLimit": {
            "type": "integer"
          }
        },
        "required": [
          "SourceFile"
        ]
      },
      "ImageConfigScriptOptions": {
        "type": "object",
        "properties": {
          "SourceFile": {
            "type": "string"
          }
        },
        "required": [
          "SourceFile"
        ]
      },
      "ImageTemplateConfig": {
        "type": "object",
        "properties": {
          "Template": {
            "type": "string"
          },
          "Enabled": {
            "type": "boolean",
            "default": true
          },
          "Connections": {
            "type": "array",
            "description": "Databases. Currently available only ['postgres']",
            "items": {
              "type": "string"
            }
          },
          "Groups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "Version": {
            "type": "string"
          },
          "ContainerOptions": {
            "$ref": "#/components/schemas/ContainerOptions"
          },
          "PackageFiles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "Workdir": {
            "type": "string"
          },
          "Provider": {
            "type": "string"
          },
          "IsSupportPackage": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "Template",
          "Groups",
          "Enabled",
          "ContainerOptions",
          "Version",
          "Workdir",
          "Provider",
          "IsSupportPackage",
          "Connections",
          "PackageFiles"
        ]
      },
      "ImageActionConfig": {
        "type": "object",
        "properties": {
          "IsDefault": {
            "type": "boolean",
            "default": false
          },
          "Id": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "CompileCmd": {
            "type": "string"
          },
          "RunCmd": {
            "type": "string"
          },
          "ScriptOptions": {
            "$ref": "#/components/schemas/ImageConfigScriptOptions"
          },
          "DefaultFiles": {
            "type": "object",
            "example": {
              "index.js": "...some content",
              "src/script.js": "...some content",
              "foo/bar/script.js": "...some content",
              "package.json": "...some content"
            },
            "additionalProperties": {
              "type": "string"
            }
          },
          "EnableExternalCommands": {
            "type": "string",
            "description": "It allows overriding CompileCmd and RunCmd in each request.",
            "enum": [
              "none",
              "run",
              "compile",
              "all"
            ],
            "default": "all"
          }
        },
        "required": [
          "Id",
          "Name",
          "CompileCmd",
          "RunCmd",
          "ScriptOptions",
          "DefaultFiles",
          "IsDefault",
          "EnableExternalCommands"
        ]
      },
      "ImageConfig": {
        "type": "object",
        "allOf": [
          {
            "type": "object",
            "properties": {
              "Actions": {
                "type": "object",
                "additionalProperties": {
                  "$ref": "#/components/schemas/ImageActionConfig"
                }
              }
            },
            "required": [
              "Actions"
            ]
          },
          {
            "$ref": "#/components/schemas/ImageTemplateConfig"
          }
        ]
      },
      "ActionItemResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ImageActionConfig"
          },
          {
            "$ref": "#/components/schemas/ImageTemplateConfig"
          }
        ]
      },
      "ActionListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ActionItemResponse"
        }
      },
      "TemplateItemResponse": {
        "type": "object",
        "properties": {
          "TemplateId": {
            "type": "string"
          },
          "Actions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}